services/forecasting/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
tensorflow==2.7.0
statsmodels==0.13.0
joblib==1.1.0
kafka-python==2.0.2
pydantic==1.8.2 